==> 05_upsetPlots_sampleSize_librarySize_qPCRload.Rmd <==
---
title: "Plot sample number in upsetPlots"
author: "Rebecca L. Knoll"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r load libraries and run functions script}
library(microViz)

pacman::p_load(tidyverse, phyloseq, magrittr, janitor, microbiome, knitr, lubridate, naniar, readxl, ggplot2, ggpubr)

library("pheatmap")
library(dendextend)
library(vegan)
library(ComplexUpset) # excellent tutorial: https://krassowski.github.io/complex-upset/articles/Examples_R.html

ps_full <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/ps_IMProveCF_patientsAndControls_Run1-23_18102023.rds")

ps_full_relab <- transform_sample_counts(ps_full, function(x) x/sum(x)*100)

#extract metadata
metadata<- as(sample_data(ps_full),"data.frame")

```

# Sputum
```{r}
metadata <- metadata %>% 
  filter(project=="IMP") 

#  DataFrame setup
count_df <-  metadata%>%
  filter(material == "Sputum") %>% 
    tabyl(id.x, visit_cal_cor) 

count_df <- clean_names(count_df) %>% 
  rename_with(~ str_replace_all(., "^x", "Visit"), starts_with("x"))
  
# plot
p1 <- ComplexUpset::upset(data=count_df, intersect= c("Visit1", "Visit2", "Visit3", "Visit4", "Visit5", "Visit6", "Visit7", "Visit8", "Visit9"), min_size=1, width_ratio=0.2, height_ratio=1, sort_sets=FALSE, stripes=c("darkgrey",  "#C3BC3FFF", "#6388B4FF", "#6388B4FF", "#6388B4FF","#BB7693FF","#BB7693FF","#55AD89FF","#55AD89FF"))
sputum <- p1 + ggtitle('Sputum samples')
sputum

#Extract plot data to find intersection coordinates
p1_data <- ggplot_build(p1)$data[[2]]  # Layer 2 usually contains the intersection points

# Add a white plus sign to the black dots
sputum <- p1 +
  geom_point(
    data = p1_data,  # Extract the data from the ComplexUpset plot
    aes(x = x, y = y),
    shape = 3,       # Shape 3 is a plus sign
    size = 1.5,        # Adjust size of the plus sign
    color = "lightgrey", # Make the plus sign white
    stroke = 1       # Adjust stroke width
  ) +
  ggtitle('Sputum samples')

sputum
```

# Throat
```{r}
metadata <- metadata %>% 
  filter(project=="IMP") 

#  DataFrame setup
count_df <-  metadata%>%  
  filter(material == "Throat")%>%
    tabyl(id.x, visit_cal_cor) 

count_df <- clean_names(count_df) %>% 
  rename_with(~ str_replace_all(., "^x", "Visit"), starts_with("x"))
  
# plot
p1 <- ComplexUpset::upset(data=count_df, intersect= c("Visit1", "Visit2", "Visit3", "Visit4", "Visit5", "Visit6", "Visit7", "Visit8", "Visit9"), min_size=1, width_ratio=0.2, height_ratio=1, sort_sets=FALSE, stripes=c("darkgrey",  "#C3BC3FFF", "#6388B4FF", "#6388B4FF", "#6388B4FF","#BB7693FF","#BB7693FF","#55AD89FF","#55AD89FF"))


#Extract plot data to find intersection coordinates
p1_data <- ggplot_build(p1)$data[[2]]  # Layer 2 usually contains the intersection points

# Add a white plus sign to the black dots
throat <- p1 +
  geom_point(
    data = p1_data,  # Extract the data from the ComplexUpset plot
    aes(x = x, y = y),
    shape = 3,       # Shape 3 is a plus sign
    size = 1.5,        # Adjust size of the plus sign
    color = "lightgrey", # Make the plus sign white
    stroke = 1       # Adjust stroke width
  ) +
  ggtitle('Throat samples')

throat
```

# Stool
```{r}
metadata <- metadata %>% 
  filter(project=="IMP") 

#  DataFrame setup
count_df <-  metadata%>%  
  filter(material == "Stool")%>%
    tabyl(id.x, visit_cal_cor) 

count_df <- clean_names(count_df) %>% 
  rename_with(~ str_replace_all(., "^x", "Visit"), starts_with("x"))
  
# plot
p1 <- ComplexUpset::upset(data=count_df, intersect= c("Visit1", "Visit2", "Visit3", "Visit4", "Visit5", "Visit6", "Visit7", "Visit8", "Visit9"), min_size=1, width_ratio=0.2, height_ratio=1, sort_sets=FALSE, stripes=c("darkgrey",  "#C3BC3FFF", "#6388B4FF", "#6388B4FF", "#6388B4FF","#BB7693FF","#BB7693FF","#55AD89FF","#55AD89FF"))
stool <- p1 + ggtitle('Stool samples')

 #Extract plot data to find intersection coordinates
p1_data <- ggplot_build(p1)$data[[2]]  # Layer 2 usually contains the intersection points

# Add a white plus sign to the black dots
stool <- p1 +
  geom_point(
    data = p1_data,  # Extract the data from the ComplexUpset plot
    aes(x = x, y = y),
    shape = 3,       # Shape 3 is a plus sign
    size = 1.5,        # Adjust size of the plus sign
    color = "lightgrey", # Make the plus sign white
    stroke = 1       # Adjust stroke width
  ) +
  ggtitle("Stool samples")

stool


```

# work on another overview
```{r, fig.height=7, fig.width=5}
visit_sum_palette <- c("darkgrey", "#C3BC3FFF", "#6388B4FF", "#BB7693FF", "#55AD89FF", "#8176AA")

# Create a lookup table for id.x and hasSputum_group
sputum_mapping <- tibble::tribble(
  ~id.x, ~hasSputum_group,
  "11", "Sputum FU",
  "12", "No Sputum",
  "13", "Sputum NFU",
  "15", "Sputum NFU",
  "16", "Sputum NFU",
  "17", "Sputum NFU",
  "18", "No Sputum",
  "19", "No Sputum",
  "20", "No Sputum",
  "21", "Sputum FU",
  "22", "No Sputum",
  "23", "No Sputum",
  "24", "Sputum FU",
  "25", "Sputum FU",
  "26", "Sputum FU",
  "27", "Sputum FU",
  "28", "No Sputum",
  "29", "Sputum FU",
  "30", "No Sputum",
  "31", "Sputum NFU",
  "32", "Sputum NFU",
  "33", "Sputum FU",
  "35", "Sputum NFU",
  "36", "No Sputum",
  "5",  "Sputum FU",
  "6",  "Sputum FU",
  "8",  "Sputum NFU",
  "9",  "Sputum FU"
)

# Merge with metadata
metadata <- metadata %>%
  left_join(sputum_mapping, by = "id.x") %>% 
  mutate(hasSputum_group = case_when(is.na(hasSputum_group)~"No Sputum", T~ hasSputum_group))

metadata$hasSputum_group <- factor(metadata$hasSputum_group, levels = c("Sputum FU", "Sputum NFU", "No Sputum"))

metadata <- metadata %>%
  mutate(label = 
    case_when(
      material == "Sputum" ~"\u25A0",  # Rectangle
      material == "Throat" ~ "\u25CF",  # Circle
      TRUE ~ "\u25B2"  # Triangle for any other material
    ))

# Ensure id.x is sorted by hasSputum_group
metadata <- metadata %>%
  arrange(hasSputum_group, id.x) %>%  # Sort first
  mutate(id.x = factor(id.x, levels = unique(id.x)))  # Convert to factor with ordered levels

overview <- metadata %>% 
  filter(visit_sum != "Control") %>% 
  ggplot(aes(x = visit_cal_9, y = id.x)) +
  geom_tile(aes(fill = visit_sum), color = "white") +
  scale_fill_manual(values = visit_sum_palette, name= "Months from treatment start", labels=c("0", "3", "6-12", "15-18", "21-24") ) +
  geom_text(data=metadata %>% filter(material == "Sputum"), 
            aes(label = label),
            color = "black", size = 3, position = position_nudge(x = -0.2)) +
  geom_text(data=metadata %>% filter(material == "Throat" & visit_sum != "Control"), 
            aes(label = label),
            color = "black", size = 3, position = position_nudge(x = 0))+    
geom_text(data=metadata %>% filter(material == "Stool" & visit_sum != "Control"), 
            aes(label = label),
            color = "black", size = 3, position = position_nudge(x = 0.2))  +
  theme_classic()+
  theme(legend.position = "bottom")+
  xlab("Visit")+
  ylab("Patient ID")+
  facet_grid(rows = vars(hasSputum_group), scales = "free", space="free_y")

overview

# Create a legend-friendly dataset
legend_data <- data.frame(
  material = c("Sputum", "Throat", "Stool"),
  x = c(1, 2, 3),
  y = c(1, 1, 1),
  shape = c(15, 16, 17), # Different shapes for different sample types
  label = c("Sputum", "Throat", "Stool")
)

legend_data$material <- factor(legend_data$material, levels = c("Sputum", "Throat", "Stool") )

# Add legend manually using an extra plot
legend_plot <- ggplot(legend_data, aes(x = x, y = y, shape = material, label = label)) +
  geom_point(size = 4, aes(shape = material)) +
  geom_text(nudge_y = 0.2, size = 4) +
  scale_shape_manual(values = c("Sputum" = 15, "Throat" = 16, "Stool" = 17), name = "Sample type") +
  theme_void() +
  theme(legend.position = "bottom")
leg_m <- get_legend(legend_plot)
leg_m <- as_ggplot(leg_m)

overview_leg <- ggarrange(overview, leg_m, nrow = 2, heights = c(9,1), labels = c("a",""))
overview_leg 
```


# plot 16S library size 
```{r}
visit_sum_palette <- c("black", "#C3BC3FFF", "#6388B4FF", "#BB7693FF", "#55AD89FF", "#8176AA")

metadata<- as(sample_data(ps_full),"data.frame")
metadata$visit_sum

metadata<-metadata %>% 
  mutate(visit_sum=case_when(visit_sum=="8-10"~"8-9", T~visit_sum))
metadata$visit_sum

visit_sum_labels <- c("0", "3", "6-12", "15-18", "21-24", "Control")
d <- metadata %>% 
  ggplot(aes(material, total_reads, fill = visit_sum)) +
  geom_boxplot(alpha = 0.65) +
  scale_fill_manual(values = visit_sum_palette, labels = visit_sum_labels) +  # Corrected part
  scale_y_log10() +
  theme_bw() +
  theme(legend.position = "none", text = element_text(size = 11)) +
  labs(x = "", y = "Library size", fill = "Months from treatment start")
d

legend <- get_legend(d+theme(legend.position = "bottom", legend.title.position = "top"))

legend2 <- get_legend(d + 
                        theme(legend.position = "bottom", legend.title.position = "left")+ 
                        guides(fill = guide_legend(nrow = 1)))

saveRDS(legend2, "/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/legend_visit_sum.rds")
as_ggplot(legend2)

summary(lmerTest::lmer(total_reads ~ visit_sum + + (1|material) + (1|id.x), data=metadata))
kruskal.test(total_reads ~ visit_sum, data=metadata)

summary(lmerTest::lmer(total_reads ~ material + (1|id.x), data=metadata))
kruskal.test(total_reads ~ material, data=metadata)
```

```{r}

e<- metadata %>% 
  ggplot(aes(material, quantity_mean, fill=visit_sum))+
  geom_boxplot(alpha=0.65) +
  scale_fill_manual(values = visit_sum_palette)+
  scale_y_log10()+
  theme_bw()+
  theme(legend.position = "none", text = element_text(size = 11))+
  labs(x="", y="16S copy number", fill="Visits")
e

summary(lmerTest::lmer(quantity_mean ~ visit_sum + (1|material) + (1|id.x), data=metadata))
kruskal.test(quantity_mean ~ visit_sum, data=metadata)

summary(lmerTest::lmer(quantity_mean ~ material + (1|id.x), data=metadata))
kruskal.test(quantity_mean ~ material, data=metadata)
```

```{r, fig.height=20, fig.width=16}
clin_evol <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/clinical_evolution.rds")

samples <- ggarrange(sputum, throat, stool, labels = c("a", "b", "c"), nrow = 1)


bio <- ggarrange(e, d, legend,labels = c("e", "f"), nrow = 1)

ggarrange(samples, clin_evol, bio, nrow = 3, ncol=1, heights = c(0.5,1.15,0.35))

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/overview.png", overview, height = 5, width = 5)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/supp_fig1_clinevol_upsetplots.png", height = 20, width = 16)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/pdf_figures/supp_fig1_clinevol_upsetplots.pdf", height = 20, width = 16)
```

# Ext. Figure 1
```{r, fig.width=12, fig.height=12}
# In order to add the overview to the revision I will split the Extended Figure 1 into 2 parts
samples_3 <- ggarrange(sputum, throat, stool, labels = c("b", "c", "d"), nrow = 3)
grid.arrange(overview_leg, samples_3, NULL, ncol = 2, nrow = 2)
bio <- ggarrange(e, d,  labels = c("e", "f"), nrow = 1)
dummy <- as_ggplot(NULL)

# Arrange the ggplot objects in a grid
grid_arranged <- grid.arrange(overview_leg, 
                              samples_3, 
                              bio,
                              legend2,
  ncol = 2, nrow = 3,
  heights = c(4, 1.25, 0.25),
  widths=c(1,1),
  layout_matrix = rbind(c(1,2),
                        c(3,2),
                        c(4,2))
)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/Revision/Rev_figures/ExtData1.png", grid_arranged, width = 12, height = 12, dpi = 600)

```
