==> 09_dominance_alluvialPlot_sputum.Rmd <==
---
title: "GG alluvial plot for genus dominance in sputum"
author: "Rebecca L. Knoll"
date: " last edit `r format(Sys.time(), '%d %B, %Y')`"
output: 
  html_document:
    code_folding: show
    number_sections: yes
    smart: no
    toc: yes
    df_print: paged
    toc_float: yes
---

```{r setup 2, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, cache= FALSE, warning = FALSE)
```

```{r run import script, include=FALSE}
# run script to tidy the data and to load packages, create phyloseq object (ps_clean) and load libraries
ps_full <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/ps_IMProveCF_patients_Run1-23_18102023.rds")

pacman::p_load(tidyverse, phyloseq, magrittr, janitor, microbiome, knitr, lubridate, naniar, readxl, ggplot2, ggpubr, ggalluvial)

source("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/IMMProveCF/functions_full.R")

# assign colors

library(paletteer)
dom_palette <- c(paletteer_d("ggthemes::Nuriel_Stone"))
dom_palette <- c(Streptococcus="#69b3a2", Staphylococcus = "#8175AA", Fusobacterium="#6FB899FF", Prevotella_7="#31A1B3FF", Rothia="#027B8EFF", Pseudomonas="#EE6AA7", Eikenella="#94D0C0FF", Haemophilus="#CCB22BFF", Achromobacter="#9F8F12FF", Gemella= "#97CFD0" , Moraxella= "#6FB899", `missing sample` = "#CFCFCF")

```

```{r subset per material}

# subset sputum
ps_full_sputum <- subset_samples(ps_full, material== "Sputum")

# remove zero abundances from  dataset
ps_sputum <- microViz::tax_filter(ps_full_sputum, min_prevalence = 1,
  prev_detection_threshold = 1, min_total_abundance = 0, min_sample_abundance = 0, tax_level = NA,
  names_only = FALSE, use_counts = TRUE, undetected = NULL,verbose = TRUE)

# calculate relative abundances
ps_full_relab <- transform_sample_counts(ps_full, function(x) x/sum(x))
ps_sputum_relab <- transform_sample_counts(ps_sputum, function(x) x/sum(x))

```

```{r find dominant genus; most abundant taxa per patient in sputum}

# Analysis on most abundant taxa per patient in sputum
ps_sputum_glom <- tax_glom(ps_sputum_relab, taxrank = "Genus")

#Get top taxa per patient
#find.top.taxa2 is sourced from functions.R
top.sputum<- find.top.taxa2(ps_sputum_glom, "Genus",1)
top.sputum$Species<- NULL

rslt <- top.sputum[, "taxa"]
dd <- matrix(unlist(rslt), nrow=1)
colnames(dd) <- rownames(top.sputum)
top.sputum <- t(dd)

top.sputum_df <- data.frame(x1 = row.names(top.sputum), top.sputum)%>%
  mutate(dominantGenus = top.sputum)
top.sputum_df$top.sputum<- NULL

##Add dominant Genus to ps_sputum_glom sample data
ps_sputum_glom <- microViz::ps_join(ps_sputum_glom, top.sputum_df, by = "x1")
```

```{r reshape metadata for ggalluvial}

# extract metadata
full_metadata <- as(sample_data(ps_sputum_glom), "data.frame")

# pivot wider per visit_sum ( reduce first number of columns for more clearness)

metadata <- full_metadata%>%
  select(x1, visit_sum, visit_cal_cor,  id.y, id_visit, dominantGenus)

m_count <- metadata%>%
  group_by(visit_sum)%>%
  count(dominantGenus)%>%
  pivot_wider(names_from = visit_sum, values_from = n, names_prefix = "V")

metadata_w <- metadata%>%
  pivot_wider(names_from = visit_cal_cor, values_from = dominantGenus, names_prefix = "V")%>%
  mutate(n_observation = c(1:57))

metadata_w%>%count(V1)

metadata_w2 <- metadata%>%
  select(-c(visit_sum, x1))%>%
  pivot_wider(names_from = visit_cal_cor, values_from = dominantGenus, names_prefix = "V")%>%
  mutate(V8V9=coalesce(V8,V9))%>%
  group_by(id.y) %>% 
  summarise(across(everything(), ~ max(., na.rm = T)))%>% # merges the rows of same id patients
  mutate(domGenus = factor(V1, levels=c("Streptococcus", "Staphylococcus", "Fusobacterium", "Prevotella_7", "Rothia", "Pseudomonas", "Eikenella", "Haemophilus", "Achromobacter","Gemella" , "Moraxella", "`missing sample`")))
  
```

```{r alluvial plot}

# display only samples with follow up
fu_id <- metadata%>%
  filter(visit_cal_cor!=1)%>%
  select(id.y)%>%
  filter(id.y!="IMP8")

fu_id <- fu_id$id.y

#in order to have each visits per patient in my dataframe, even if they had no sampling conducted at that timepoint, I need to add a dummie row for it

visit_fictional <- c(1:9)

vist_fic_df <- merge(fu_id, visit_fictional)%>%
 mutate(id_visit =  paste(x, y, sep="V"))%>%
  distinct(id_visit, .keep_all = T)

metadata_fic <- metadata%>%
  filter(id.y%in%fu_id)%>%
  right_join(vist_fic_df, by="id_visit")%>%
  mutate(dominantGenus_new = as_factor(case_when(is.na(dominantGenus) ~ "missing sample")))%>%
  mutate(dominantGenus_new = as_factor(coalesce(dominantGenus, dominantGenus_new)))%>%
  mutate(visit=as_factor(y))%>%
  mutate(visit_small = as_factor(case_when(visit=="9"~"8", TRUE~visit)))

 metadata_fic$visit_small <-  fct_relevel(metadata_fic$visit_small, sort)

# for alluvial format I now have to remove the "double" visit 8

id_rm <- c("IMP29V9","IMP5V9", "IMP6V8", "IMP8V8", "IMP21V9", "IMP25V9", "IMP11V8", "IMP26V9", "IMP24V9", "IMP27V9", "IMP33V9", "IMP9V9")

dom_spu <- metadata_fic%>%
  filter(!id_visit%in%id_rm)%>%
  ggplot(aes(x = visit_small, stratum = dominantGenus_new, alluvium = x,
           fill =dominantGenus_new)) +
  scale_fill_manual(values= dom_palette,guide =
                         guide_legend(label.theme = element_text(angle = 0, face = "italic", size = 16)))+
  geom_lode() + 
  geom_flow(curve_type = "arctangent",width = .3, color="black") +
  geom_stratum(alpha = 0.6)+
  theme_pubr()+
  theme(legend.position="right", text = element_text(size=20))+
  labs(fill="Dominant genus")+
  labs(x="Months from treatment start")+
  labs(y = "n samples")+
  scale_x_discrete(labels=c("0", "3","6","9","12","15","18","21-24"))
dom_spu

saveRDS(dom_spu, "/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/alluvial_dominantGenus.rds")
#ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/alluvial_dominantGenus.pdf", dom_spu, width = 12, height = 7)
```
