==> 25a_combine_sputum_figure.Rmd <==
---
title: "Combine sputum plot"
author: "Rebecca L. Knoll"
date: "`r Sys.Date()`"
output: html_document
---

```{r run import script, include=FALSE}
# run script to tidy the data and to load packages, create phyloseq object (ps_clean) and load libraries
#install.packages("ggplot2")
#install.packages("gridExtra")
pacman::p_load(tidyverse, phyloseq, magrittr, janitor, microbiome, microViz, knitr, lubridate, naniar, readxl, ggplot2, ggpubr, ggalluvial, metadeconfoundR)

#pacman::p_load(rmarkdown, tidyverse, phyloseq, magrittr, janitor, microbiome, knitr, lubridate, naniar, readxl, ggplot2, ggpubr,rstatix, metadeconfoundR, microViz)

library(gridExtra)
library(scales)
```

# read figures from rds files
```{r}
dom_spu <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/alluvial_dominantGenus.rds")
tax_sputum <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/da_sputum.rds")
tax_sputum_woLegend <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/da_sputum_woLegend.rds")
metad_sputum <-  readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/sputum_metaD_25percentprevalence.rds")
staph <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/cor_staph_chloride_sputum.rds")
hm_div_sputum <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/sputum_div_hm.rds")
star <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/starplot_sputum.rds")
permanova_bubble <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/PERMANOVA_timepoints_sputum.rds")
permanova_univariate <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/sputum_univariate_permanova_stratified.rds")
permanova_combined<- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/sputum_combined_permanova_stratified_31Samples:withoutETI.rds")
permanova_barplot_2 <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/sputum_combined_permanova_barplot.rds")

sputum_shannon<- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/sputum_shannon.rds")
sputum_observed <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/sputum_observed.rds")

sputum_barplot <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/sputum_barplot.rds")

sputum_lolli <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/sputum_permanova_lollipop_plot.rds")
```

# upload pdf/png files
```{r}
library(png)
img <- readPNG("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/figures/sweat_chloride_mediation_crop.png")

# Convert it to a grob
library(grid)
g <- rasterGrob(img)

# Use gridExtra to arrange it
library(gridExtra)
g <- grid.arrange(g)
g <- as_ggplot(g)
```

```{r}
ic25 <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/figures/ic25_figure_sputum.rds")
```

# harmonize font size
```{r}
dom_palette <- c(
  Streptococcus = "#69b3a2", 
  Staphylococcus = "#E55E00", 
  Fusobacterium = "#6FB899FF", 
  Prevotella_7 = "#31A1B3FF", 
  Rothia = "#027B8EFF", 
  Pseudomonas = "#EE6AA7", 
  Eikenella = "#94D0C0FF", 
  Haemophilus = "#CCB22BFF", 
  Achromobacter = "#9F8F14FF", 
  Gemella = "#97CFD0", 
  Moraxella = "#6FB899", 
  `missing sample` = "#CFCFCF"
)

dom_palette_labels <- c(
  Streptococcus = "Streptococcus",
  Staphylococcus = "Staphylococcus",
  Fusobacterium = "Fusobacterium",
  Prevotella_7 = "Prevotella 7",
  Rothia = "Rothia",
  Pseudomonas = "Pseudomonas",
  Eikenella = "Eikenella",
  Haemophilus = "Haemophilus",
  Achromobacter = "Achromobacter",
  Gemella = "Gemella",
  Moraxella = "Moraxella",
  `missing sample` = "no sample"
)

dom_spu <- dom_spu +
  theme(text = element_text(size = 14), 
        legend.text = element_text(size = 14), legend.title.position = "top") +
  guides(fill = guide_legend(nrow = 4, position = "top", 
                             label.theme = element_text(angle = 0, face = "italic", size = 14))) +
  scale_fill_manual(values = dom_palette, labels= dom_palette_labels)


tax_sputum <- tax_sputum+theme(legend.text = element_text(size = 12), axis.text.x = element_text(size = 14),axis.text.y = element_text(size = 14))+theme_pubclean(base_size = 14)

star <- star + theme(axis.text= element_text(size = 14), legend.title.position = "top", legend.position = "top", axis.title = element_text(size=14), legend.text = element_text(size = 14), legend.title =  element_text(size = 14))

sputum_lolli <- sputum_lolli+ theme(axis.text= element_text(size = 14), axis.title = element_text(size=14), legend.text = element_text(size = 14), legend.title =  element_text(size = 14), legend.title.position = "left",  legend.position = "top")+ 
 guides(color=guide_legend(nrow = 2, position = "top")) 

metad_sputum <- metad_sputum + theme(axis.text.x=element_text(size=14, angle = 40, hjust = 1, vjust = 1.05), axis.text.y=element_text(size=14, face="italic"), legend.title = element_text(size=14), legend.position = "bottom")
  
hm_div_sputum <- hm_div_sputum + theme(legend.position = "bottom", text = element_text(size = 14), legend.box = "horizontal") + scale_fill_gradient2(name= "Effect size", low = "#746999", high = "#69b3a2", mid = "white", midpoint = 0, guide = guide_colorbar(raster = F),labels = scales::number_format(accuracy = 0.1))+
  guides(title.position = "left")

staph <- staph +theme(text = element_text(size = 14),legend.text = element_text(size = 14), axis.text = element_text(size = 14), axis.title = element_text(size=14), legend.title =  element_text(size = 14), legend.position = "right")+ guides(color=guide_legend(nrow = 5, title = "Month from \ntreatment start")) 

sputum_observed <- sputum_observed+theme(text = element_text(size = 14))
sputum_shannon <- sputum_shannon+theme(text = element_text(size = 14))
sputum_barplot <- sputum_barplot+theme(text = element_text(size = 14))

permanova_bubble <- permanova_bubble + theme(legend.position = "top", legend.title.position = "top", text = element_text(size = 14))

ic25 <- ic25+theme(text = element_text(size = 14))

```

```{r, fig.width=16.5, fig.height=14}
library(ggplot2)
library(scales)

grid_arranged <- grid.arrange(
  star+ggtitle("") + labs(tag = "a"), #1
  permanova_bubble +labs(tag = "b"), #2
  dom_spu+ labs(tag = "c"), #3
  sputum_lolli + labs(tag = "d"), #4
  staph+ labs(tag = "e"), #5
  g + labs(tag = "f"), #6
  ic25 + labs(tag = "g"), #7
 
  ncol = 3, nrow = 3,
  heights = c(1.2, 1, 1),
  widths=c(1,1,1),
  layout_matrix = rbind(c(1,2,3),
                        c(4,5,6),
                        c(7,7,7))
                       )

# Print or save the arranged plot
print(grid_arranged)
ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/sputum_figure_12032024.png",grid_arranged, dpi = 300, width = 16.5, height = 14)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/sputum_figure_12032024.pdf",grid_arranged, dpi = 300, width = 16.5, height = 14)
```

# Supplementary figure for sputum
```{r, fig.height=8, fig.width=8}
grid_arranged <- grid.arrange(
  sputum_shannon + labs(tag = "a") ,
  sputum_observed + labs(tag = "b") ,
  tax_sputum_woLegend + labs(tag = "c"), #+ggtitle(""),
  hm_div_sputum + labs(tag = "d"),
  ncol = 2, nrow = 3,
  heights = c(1, 1, 1.5),
  widths=c(1,1.75),
  layout_matrix = rbind(c(1,3),
                        c(2,3),
                        c(4,3))
                       )

```


```{r, fig.height=18, fig.width=16.5}
grid_arranged <- grid.arrange(
  sputum_shannon + labs(tag = "a") ,
  sputum_observed + labs(tag = "b") ,
  metad_sputum + labs(tag = "c") +ggtitle(""),
  tax_sputum + labs(tag = "d")+theme(legend.position = "none"), #+ggtitle(""),
  hm_div_sputum + labs(tag = "e") + theme(legend.position = "bottom"),
  sputum_barplot+theme(legend.position = "bottom"), 
  ncol = 3, nrow = 4,
  heights = c(0.75, 1.2, 1, 1),
  widths=c(1,1,2.5),
  layout_matrix = rbind(c(1,2,3),
                        c(4,4,3),
                        c(5,5,3),
                        c(6,6,6))
                       )
ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/041224_sputum_figure_supp.png",grid_arranged, dpi = 300, width = 16, height = 18)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/Revision/Rev_figures/supp_figure2.pdf",grid_arranged, dpi = 300, width = 16, height = 18)
```
