==> 25b_combine_stool_figure.Rmd <==
---
title: "stool figure"
author: "Rebecca L. Knoll"
date: "`r Sys.Date()`"
output: html_document
---

```{r run import script, include=FALSE}
# run script to tidy the data and to load packages, create phyloseq object (ps_clean) and load libraries
#install.packages("ggplot2")
#install.packages("gridExtra")
pacman::p_load(tidyverse, phyloseq, magrittr, janitor, microbiome, microViz, knitr, lubridate, naniar, readxl, ggplot2, ggpubr, ggalluvial, metadeconfoundR)

#pacman::p_load(rmarkdown, tidyverse, phyloseq, magrittr, janitor, microbiome, knitr, lubridate, naniar, readxl, ggplot2, ggpubr,rstatix, metadeconfoundR, microViz)

library(gridExtra)
library(scales)
library(ggplot2)
library(metadeconfoundR)

packageVersion("ggplot2")
library(dplyr)
library(gridExtra)
library(scales)
#remotes::install_version("ggplot2", version = "3.1.0")
library(ggplot2)
```

# read figures from rds files
```{r}
tax_stool <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/relAb_cuneiform_plot_stool_patients.rds")
tax_stool

tax_stool_legend <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/da_stool.rds")
tax_stool_legend

metad_stool <-  readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/stool_metaD_25percentprevalence.rds")
metad_stool

cor <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/cor_esch_cale_stool.rds")
cor

hm_div_stool <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/stool_div_hm.rds")
hm_div_stool

star <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/starplot_stool.rds")
star <- star+ggtitle("")+theme(text = element_text(size = 14))

stool_shannon <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/stool_shannon.rds")
stool_observed <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/stool_observed.rds")

bubble<-  readRDS("/Users/rebecca//Documents/Forschung/IMMProveCF/R_analysis/rds_files/PERMANOVA_timepoints_stool_woControls.rds")

stool_combined_permanova_barplot <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/stool_combined_permanova_barplot.rds")

lolli <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/stool_permanova_lollipop_plot.rds")

atbx_obs<- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/atbx_obs.rds")

boxplot_ecoli <- readRDS("figures/E_coli_boxplot.rds")

ic25 <- readRDS("figures/ic25_figure_gut.rds")
```

# harmonize font size
```{r}
metad_stool <- metad_stool+ theme(axis.text.x=element_text(size=14, angle = 40, hjust = 1, vjust = 1.05), axis.text.y=element_text(size=14, face="italic"), legend.title = element_text(size=14), legend.position = "bottom")

hm_div_stool <- hm_div_stool + 
  scale_fill_gradient2(name= "Effect size\n(Cliff's delta)", low = "#746999", high = "#69b3a2", mid = "white", midpoint = 0, guide = guide_colorbar(raster = F))+  # Set the title here
  theme(
    axis.text.x = element_text(size = 14, angle = 40, hjust = 1, vjust = 1.05), 
    axis.text.y = element_text(size = 14), 
    legend.title = element_text(size = 14),  # Style the title here
    legend.position = "bottom"
  )


cor <- cor + 
theme(axis.text.x=element_text(size=14, angle = 40, hjust = 1, vjust = 1.05), 
      legend.title = element_text(size=14), legend.position = "bottom")+
  guides(color=guide_legend(nrow = 2, title = "Months from \ntreatment start")) 

bubble <- bubble+ theme(legend.position = "top", legend.title.position = "top") +  guides(fill = guide_legend(nrow = 2, position = "top",override.aes = list(size = 4)),size = guide_legend(nrow = 2, position = "top"))

lolli <- lolli+guides(color=guide_legend(nrow = 2, position = "bottom")) 
ic25 <- ic25+guides(fill=guide_legend(nrow = 4, position = "right")) 
```

# arrange figure
```{r, fig.width=16, fig.height=12}
grid_arranged <- grid.arrange(
  star+  labs(tag = "a") ,
  bubble + theme(legend.position = "none")+ggtitle("X\nX\nX\nX")+labs(tag = "b") ,
  boxplot_ecoli+ labs(tag = "c"),
  lolli+ labs(tag = "d"),
  cor+ labs(tag = "f"),
  hm_div_stool+ labs(tag = "e"),
  ic25,
  ncol = 3, nrow = 3,
  heights = c(1.1, 1.1,0.9),
  widths=c(1.2,1,0.9),
  layout_matrix = rbind(c(1,2,3),
                        c(4,6,5),
                        c(4,7,7)
                       ))

# Print or save the arranged plot
print(grid_arranged)
ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/pdf_figures/Figure3.png",grid_arranged, dpi = 600, width = 16, height = 12)
```


# Supplementary figure
```{r, fig.width=16, fig.height=10}
# Add an empty grob to create space

grid_arranged <- grid.arrange(
  stool_shannon + labs(tag = "a") ,
 stool_observed + labs(tag = "b") ,
 metad_stool+ labs(tag = "c") ,
  tax_stool_legend+ labs(tag = "d"),
  ncol = 3, nrow = 2,
  heights = c(1 , 1.5),
  widths=c(0.5,0.5,1),
  layout_matrix = rbind(c(1,2,3),
                        c(4,4,3))
                       )

# Print or save the arranged plot
print(grid_arranged)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/stool_supplement_figure.png",grid_arranged, dpi = 300, width = 16, height = 10)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/Revision/Rev_figures/supp_figure5.png",grid_arranged, dpi = 300, width = 16, height = 10)
```
