==> 25c_combine_throat_figure.Rmd <==
---
title: "Combine throat plot"
author: "Rebecca L. Knoll"
date: "`r Sys.Date()`"
output: html_document
---

```{r run import script, include=FALSE}
# run script to tidy the data and to load packages, create phyloseq object (ps_clean) and load libraries
#install.packages("ggplot2")
#install.packages("gridExtra")
pacman::p_load(tidyverse, phyloseq, magrittr, janitor, microbiome, microViz, knitr, lubridate, naniar, readxl, ggplot2, ggpubr, ggalluvial, metadeconfoundR)

#pacman::p_load(rmarkdown, tidyverse, phyloseq, magrittr, janitor, microbiome, knitr, lubridate, naniar, readxl, ggplot2, ggpubr,rstatix, metadeconfoundR, microViz)

library(gridExtra)
```

# read figures from rds files
```{r}
star <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/starplot_throat.rds")
throat_shannon<- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/throat_shannon.rds")
throat_observed<- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/throat_observed.rds")
bubble <- readRDS("~/Documents/Forschung/IMMProveCF/R_analysis/figures/PERMANOVA_timepoints_throat_woControls.rds")
perm <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/figures/sputum_univariate_permanova_stratified_vertical.rds")
throat_div_hm <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/throat_div_hm.rds")
metad <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/IMMProveCF_public/outputs/throat_metaD_25percentprevalence.rds")
BCdist <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/figures/Correlation_BCdist_intraID_spuThr.rds")

lolli <- readRDS("/Users/rebecca/Documents/Forschung/IMMProveCF/R_analysis/rds_files/throat_permanova_lollipop_plot.rds")
```

# harmonize font size
```{r}
star <- star+theme(text = element_text(size = 14))+xlim(-0.34, 0.4)
bubble <- bubble+theme(text = element_text(size = 14), legend.position = "top", legend.title.position = "top") + guides(fill = guide_legend(nrow = 2), size=guide_legend(nrow = 2))
BCdist <- BCdist+theme(text = element_text(size = 14), axis.text.x = element_text(size = 14), axis.text.y = element_text(size = 14),  legend.title.position = "top", legend.position = "top")

metad <- metad+theme(text = element_text(size = 14), axis.text.x = element_text(size = 14), axis.text.y = element_text(size = 14, face = "italic"), legend.title  = element_text(size = 14),  legend.position = "bottom")

metad <- metad +
  scale_y_discrete(position = "right")

throat_div_hm <- throat_div_hm+theme(text = element_text(size = 14),axis.text.x = element_text(size = 14), axis.text.y = element_text(size = 14), legend.position = "bottom", legend.title.position = "left")
throat_observed <- throat_observed+theme(text = element_text(size = 14))
throat_shannon<- throat_shannon+theme(text = element_text(size = 14))

perm <- perm+theme(text = element_text(size = 14))

lolli <- lolli+theme(legend.position = "top")+ guides(color = guide_legend(nrow = 2))
```

```{r, fig.width=14, fig.height=5}
grid_arranged <- grid.arrange(
  star+ggtitle("") + labs(tag = "a"), #1
  bubble + theme(legend.position = "none")+ggtitle("X\nX\nX\nX")+labs(tag = "b"), #2
  BCdist + labs(tag = "c"), #3
  ncol = 3, nrow = 1,
  #heights = c(1.2, 1.2, 1, 1, 1),
  widths=c(1.1,1,0.85),
  layout_matrix = rbind(c(1,2,3))
                       )

# Print or save the arranged plot
print(grid_arranged)
ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/throat_figure.png",grid_arranged, dpi = 300, width = 14.5, height = 5)
ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/Revision/Rev_figures/figure4.pdf",grid_arranged, dpi = 300, width = 14.5, height = 5)
```

# Supplementary figure
```{r, fig.width=16.5, fig.height=10}
# Add an empty grob to create space
spacer <- nullGrob()

grid_arranged <- grid.arrange(
  throat_shannon + labs(tag = "a"),
  throat_observed + labs(tag = "b"),
  lolli + labs(tag = "c"),
  throat_div_hm + labs(tag = "d"),
  spacer,  # Spacer added here to make space
  metad+ labs(tag = "e"),
  ncol = 4, nrow = 2,
  layout_matrix = rbind(c(1, 3, 5, 6),
                        c(2, 4, 5, 6)),
  widths = c(0.75, 1.05, 0.1, 1.6)  # adjust "0.3" to control space width
)

# Print or save the arranged plot
print(grid_arranged)

ggsave("/Users/rebecca/Documents/Forschung/IMMProveCF/Manuscript_16S/Nature_Communications/Revision/Rev_figures/ExtDat6.pdf",grid_arranged, dpi = 300, width = 16, height = 10)
```
